requirements.txt
flask
requests
websocket-client
google-api-python-client
google-auth-oauthlib
Pillow
icrawler
huggingface_hub